cbc_mac_breaker/message_forgery2/frontend/requirements.txt
Werkzeug==2.0.1
Flask==2.0.1
PyCryptodome==3.11.0
requests==2.26.0
